Initial commit
Modifies the standard Foundry circle template to support using the width
field to indicate the size of a center square that the circular template
is moved around. i.e. to indicate reach or an aura with a circular
radius centered on a creature

Does not currently update the template ruler posision to account for the
added width. There is currently no way to place a circular template with
the width field set with an Item or through the GUI.